Platformer demo + replication
I created a small platformer demo in order to test the replication. There seems to be a slight latency between the client and the platforms making its movement slightly buggy. The dynamic object like cubes from the map, while taking the right impulse when interacted with, the client and server perspective on them seems to be a bit off after they are pushed around.